Move products controller logic into own file
